Switch to fasttext based language detection
FastText based language detection model outperforms Lingua for
some selected texts. Lingua is retained as a fallback model, as
the FastText based model may emit a prediction that is not in
the set of supported languages, while Lingua can be configured
to constrain the set of predicted languages, which allows for
increased accuracy in its own way.